Created new glue.utils sub-package for general utilities that are not related to data linking (e.g. WCS, matplotlib, etc.)
