+++ glue/clients/image_client.py
@@ -8,16 +8,15 @@ import numpy as np
 from ..external.modest_image import extract_matched_slices
 from ..core.exceptions import IncompatibleAttribute
 from ..core.data import Data
-from ..core.util import lookup_class
 from ..core.subset import Subset, RoiSubsetState
 from ..core.roi import PolygonalROI
 from ..core.message import ComponentReplacedMessage
 from ..core.callback_property import (
     callback_property, CallbackProperty)
 from ..core.edit_subset_mode import EditSubsetMode
+from ..utils import lookup_class, defer_draw
 
 from .viz_client import VizClient, init_mpl
-from .util import defer_draw
 from .layer_artist import (ScatterLayerArtist, LayerArtistContainer,
                            ImageLayerArtist, SubsetImageLayerArtist,
                            RGBImageLayerArtist,
