+++ glue/clients/scatter_client.py
@@ -8,12 +8,13 @@ from ..core.client import Client
 from ..core.data import Data, IncompatibleAttribute, ComponentID, CategoricalComponent
 from ..core.subset import RoiSubsetState, RangeSubsetState
 from ..core.roi import PolygonalROI, RangeROI
-from ..core.util import relim, lookup_class
+from ..core.util import relim
 from ..core.edit_subset_mode import EditSubsetMode
 from ..core.message import ComponentReplacedMessage
+from ..utils import lookup_class, visible_limits
 from .viz_client import init_mpl
 from .layer_artist import ScatterLayerArtist, LayerArtistContainer
-from .util import visible_limits, update_ticks
+from .util import update_ticks
 from ..core.callback_property import (CallbackProperty, add_callback,
                                       delay_callback)
