+++ glue/core/util.py
@@ -12,15 +12,6 @@ from ..external.six.moves import reduce
 from ..external.six import string_types
 
 
-def shape_to_string(shape):
-    """
-    On Windows, shape tuples use long ints which results in formatted shapes
-    such as (2L, 3L). This function ensures that the shape is always formatted
-    without the Ls.
-    """
-    return "({0})".format(", ".join(str(int(item)) for item in shape))
-
-
 def identity(x):
     return x
 
@@ -46,37 +37,6 @@ def file_format(filename):
     else:
         result = filename.lower().rsplit('.', 1)[1]
     return result
-
-
-def point_contour(x, y, data):
-    """Calculate the contour that passes through (x,y) in data
-
-    :param x: x location
-    :param y: y location
-    :param data: 2D image
-    :type data: :class:`numpy.ndarray`
-
-    Returns:
-
-       * A (nrow, 2column) numpy array. The two columns give the x and
-         y locations of the contour vertices
-    """
-    try:
-        from scipy import ndimage
-    except ImportError:
-        raise ImportError("Image processing in Glue requires SciPy")
-
-    inten = data[y, x]
-    labeled, nr_objects = ndimage.label(data >= inten)
-    z = data * (labeled == labeled[y, x])
-    y, x = np.mgrid[0:data.shape[0], 0:data.shape[1]]
-    from matplotlib import _cntr
-    cnt = _cntr.Cntr(x, y, z)
-    xy = cnt.trace(inten)
-    if not xy:
-        return None
-    xy = xy[0]
-    return xy
 
 
 def split_component_view(arg):
@@ -124,42 +84,6 @@ def join_component_view(component, view):
         result = [component, view]
 
     return tuple(result)
-
-
-def view_shape(shape, view):
-    """Return the shape of a view of an array
-
-    :param shape: Tuple describing shape of the array
-    :param view: View object -- a valid index into a numpy array, or None
-
-    Returns equivalent of np.zeros(shape)[view].shape
-    """
-    if view is None:
-        return shape
-    shp = tuple(slice(0, s, 1) for s in shape)
-    xy = np.broadcast_arrays(*np.ogrid[shp])
-    assert xy[0].shape == shape
-
-    return xy[0][view].shape
-
-
-def stack_view(shape, *views):
-    shp = tuple(slice(0, s, 1) for s in shape)
-    result = np.broadcast_arrays(*np.ogrid[shp])
-    for v in views:
-        if isinstance(v, string_types) and v == 'transpose':
-            result = [r.T for r in result]
-            continue
-
-        result = [r[v] for r in result]
-
-    return tuple(result)
-
-
-def color2rgb(color):
-    from matplotlib.colors import ColorConverter
-    result = ColorConverter().to_rgb(color)
-    return result
 
 
 def facet_subsets(data_collection, cid, lo=None, hi=None, steps=5,
@@ -282,58 +206,6 @@ def colorize_subsets(subsets, cmap, lo=0, hi=1):
         subset.style.color = '#%2.2x%2.2x%2.2x' % (r, g, b)
 
 
-def coerce_numeric(arr):
-    """Coerce an array into a numeric array, replacing
-       non-numeric elements with nans.
-
-       If the array is already a numeric type, it is returned
-       unchanged
-
-       :param arr: array to coerce
-       :type arr: :class:`numpy.ndarray`
-
-       :returns: array.
-    """
-    # already numeric type
-    if np.issubdtype(arr.dtype, np.number):
-        return arr
-
-    if np.issubdtype(arr.dtype, np.bool_):
-        return arr.astype(np.int)
-
-    # a string dtype, or anything else
-    return pd.Series(arr).convert_objects(convert_numeric=True).values
-
-
-def check_sorted(array):
-    """ Return True if the array is sorted, False otherwise.
-    """
-    # this ignores NANs, and does the right thing if nans
-    # are concentrated at beginning or end of array
-    # otherwise, it will miss things at nan/finite boundaries
-    return not (array[:-1] > array[1:]).any()
-
-
-def lookup_class(ref):
-    """ Look up an object via it's module string (e.g., 'glue.core.Data')
-
-    :param ref: reference
-    :type ref: str
-    :rtype: object, or None if not found
-    """
-    mod = ref.split('.')[0]
-    try:
-        result = __import__(mod)
-    except ImportError:
-        return None
-    try:
-        for attr in ref.split('.')[1:]:
-            result = getattr(result, attr)
-        return result
-    except AttributeError:
-        return None
-
-
 class PropertySetMixin(object):
 
     """An object that provides a set of properties that
@@ -385,12 +257,6 @@ class CallbackMixin(object):
     def notify(self, *args, **kwargs):
         for func in self._callbacks:
             func(*args, **kwargs)
-
-
-def as_list(x):
-    if isinstance(x, list):
-        return x
-    return [x]
 
 
 class Pointer(object):
@@ -445,20 +311,6 @@ def defer(instance, method):
             orig(*a, **k)
 
 
-def as_variable_name(x):
-    """
-    Convert a string to a legal python variable name
-
-    :param x: A string to (possibly) rename
-    :returns: A legal python variable name
-    """
-    allowed = string.ascii_letters + string.digits + '_'
-    result = [letter if letter in allowed else '_' for letter in x or 'x']
-    if result[0] in string.digits:
-        result.insert(0, '_')
-    return ''.join(result)
-
-
 def disambiguate(label, taken):
     """If necessary, add a suffix to label to avoid name conflicts
 
@@ -476,72 +328,6 @@ def disambiguate(label, taken):
         candidate = label + (suffix % i)
         if candidate not in taken:
             return candidate
-
-
-def nonpartial(func, *args, **kwargs):
-    """Like functools.partial, this returns a function which,
-    when called, calls func(*args, **kwargs). Unlike functools.partial,
-    extra arguments passed to the returned function are *not* passed
-    to the input function.
-
-    This is used when connecting slots to QAction.triggered signals,
-    which appear to have different signatures, which seem to add
-    and extra argument in PyQt4 but not PySide
-    """
-    def result(*a, **k):
-        return func(*args, **kwargs)
-
-    return result
-
-
-def all_artists(fig):
-    """
-    Build a set of all Matplotlib artists in a Figure
-    """
-    return set(item
-               for axes in fig.axes
-               for container in [axes.collections, axes.patches, axes.lines,
-                                 axes.texts, axes.artists, axes.images]
-               for item in container)
-
-
-def new_artists(fig, old_artists):
-    """
-    Find the newly-added artists in a figure
-
-    :param fig: Matplotlib figure
-    :param old_artists: Return value from :func:all_artists
-    :returns: All artists added since all_artists was called
-    """
-    return all_artists(fig) - old_artists
-
-
-def remove_artists(artists):
-    """
-    Remove a collection of matplotlib artists from a scene
-
-    :param artists: Container of artists
-    """
-    for a in artists:
-        try:
-            a.remove()
-        except ValueError:  # already removed
-            pass
-
-
-def unique(array):
-    """
-    Return the unique elements of the array U, as well as
-    the index array I such that U[I] == array
-
-    :param array: The array to use
-    :returns: U, I
-    :rtype: tuple of arrays
-    """
-    # numpy.unique doesn't handle mixed-types on python3,
-    # so we use pandas
-    U, I = pd.factorize(array, sort=True)
-    return I, U
 
 
 def row_lookup(data, categories):
