+++ glue/utils/__init__.py
@@ -1,0 +1,12 @@
+"""
+General utilities not specifically related to data linking (e.g. WCS or
+matplotlib helper functions).
+
+Utilities here cannot import from anywhere else in glue, and can only import
+standard library or external dependencies.
+"""
+
+from .array import *
+from .matplotlib import *
+from .misc import *
+from .wcs import *
