+++ glue/utils/tests/test_matplotlib.py
@@ -1,0 +1,34 @@
+import numpy as np
+from numpy.testing import assert_allclose
+
+from ...tests.helpers import requires_scipy
+
+from ..matplotlib import point_contour, fast_limits
+
+
+@requires_scipy
+class TestPointContour(object):
+
+    def test(self):
+        data = np.array([[0, 0, 0, 0],
+                         [0, 2, 3, 0],
+                         [0, 4, 2, 0],
+                         [0, 0, 0, 0]])
+        xy = point_contour(2, 2, data)
+        x = np.array([2., 2. + 1. / 3., 2., 2., 1, .5, 1, 1, 2])
+        y = np.array([2. / 3., 1., 2., 2., 2.5, 2., 1., 1., 2. / 3])
+
+        np.testing.assert_array_almost_equal(xy[:, 0], x)
+        np.testing.assert_array_almost_equal(xy[:, 1], y)
+
+
+
+
+def test_fast_limits_nans():
+    x = np.zeros((10, 10)) * np.nan
+    assert_allclose(fast_limits(x, 0, 1), [0, 1])
+
+
+def test_single_value():
+    x = np.array([1])
+    assert_allclose(fast_limits(x, 5., 95.), [1, 1])
